requirements.txt
xarray
rasterio
ecmwf-opendata